rx_dma_trans_buffer [OPT]: create new buffering mechanism
The previous buffer did not buffer data when the DST_RDY from the
receiver behind this buffer has been set to 0. That was the reason for
development of this optimization which ensures that data for the new
transaction are always ready even if the next component is not ready to
accept data. This has replaced the old implementation entirely and works
by default.

The optimization has been pushed even furter by adding a one-stage skid
buffer to the output of the buffer that keeps the DST_RDY internally
asserted if the output of the skid buffer does not have valid data. This
skid buffer can be deactivated with the REG_OUT_EN generic parameter
since it adds a latency of one clock period.