tx_dma_pcie_trans_buffer [BUGFIX]: correct select of address for read/write
This swas previously based on reading the first bit that determined the
selection of address for read/write in the specific port. However the
byte enable signal from which this bit has been read is already shifted.
On the shifted signal, the first bit necessary does not equal to 1 which
creates an error when read address would be preferred instead.