property.sv [MAINTENANCE]: remove unnecessary property for output packets
The property was copied from the RX DMA which controlled that packets
come without internal gaps from the core. This makes sense if these
packets are sent towards the PCIE IP but it does not make sense in TX
DMA where data are send towards user logic. This uses MFB that supports
gaps inside a packet.