DMA CALYPTE [FEATURE]: update documentation
The update reflects the recent changes in the TX DMA like 2-region
configuration. The register map has been added to both of the
controllers (RX and TX). A documentation comment from the entity
declaration has been moved to each component's respective .rst file so
the documentation remains in one place when read from the source files.

The code coverage summary and latency summary has been removed since it
needs to be updated with regards to the recent changes in design.