tx_dma_pkt_dispatcher [MAINTENANCE]: remove dropping of packets when channel is disabled
This was removed because it is incompatible with the verification where drop is
only read on the beginning of the pipeline, e.g.
TX_DMA_CHAN_START_STOP_CTRL. However, it has been found in authors mind
that this change actually came out of the hardware testing but its cause
ended up in a brain fog.