rx_dma_calypte [MAINTENANCE]: make PCIE_HDR MVB interfaces independent
Previously, there was a hybrid of two MVB interfaces (two SRC_RDY
signals) with common data bus where data were multiplexed. This, however
creates a confusing control mechanism in the HDR_INSERTOR as well as in
HDR_MANAGER so it was removed in order to simplify the interface for
passing the PCIe headers.